+++ cluster/pulumi/validator-runbook/src/utils.ts
@@ -1,4 +1,4 @@
-import { requireEnv } from 'cn-pulumi-common';
+import { envFlag, requireEnv } from 'cn-pulumi-common';
 
 export const CLUSTER_BASENAME = requireEnv(
   'GCP_CLUSTER_BASENAME',
@@ -6,3 +6,8 @@ export const CLUSTER_BASENAME = requireEnv(
 );
 
 export const VALIDATOR_NAMESPACE = process.env.VALIDATOR_NAMESPACE || 'validator';
+
+export const VALIDATOR_PARTY_HINT = process.env.VALIDATOR_PARTY_HINT;
+export const VALIDATOR_MIGRATE_PARTY = envFlag('VALIDATOR_MIGRATE_PARTY', false);
+
+export const VALIDATOR_NEW_PARTICIPANT_ID = process.env.VALIDATOR_NEW_PARTICIPANT_ID;